Mac ARM64 integration (#19571)
* Update Mac platform to support multi-arch detection
* New Mac arm64 specific cmake files
* Remove EditorCommon.Headers causing issues with Qt6 moving and Mac

* Fix QAbstractNativeEventFilter::nativeEventFilter Signature
* Update deprecated constructor call to QMouseEvent
* Update deprecated QHoverEvent->pos() to ->position)
* Fix issue where file that needs mocing 'QtEditorApplication_mac.h' was not in any cmake file list, thus not being auto-moc'd
* Removed unneeded include for that moc
* Added static_cast<int> in numerous places to fix error "Implicit conversion loses integer precision: 'qsizetype' (aka 'long long') to 'int'"



Unit test fixes and review feedback

various compilation fixes


Mac support

+++ Code/Editor/AssetImporter/AssetImporterManager/AssetImporterManager.cpp
@@ -291,7 +291,7 @@ bool AssetImporterManager::OnKeepBothFiles(QString relativePath, QString oldAbso
     QString fileName = info.baseName(); //file name without extension
 
     int number = 1;
-    int index = destinationAbsolutePath.indexOf(extension);
+    int index = static_cast<int>(destinationAbsolutePath.indexOf(extension));
 
     QString newFileName = CreateFileNameWithNumber(number, fileName, index, extension);
     QString newDestinationAbsolutePath = subPath + '/' + newFileName;
@@ -432,7 +432,7 @@ bool AssetImporterManager::ProcessFileMethod(ProcessFilesMethod processMethod, Q
 
 void AssetImporterManager::ProcessCopyFiles()
 {
-    int numberOfFiles = m_pathMap.size();
+    int numberOfFiles = static_cast<int>(m_pathMap.size());
     int numberOfProcessedFiles = 0;
 
     ProcessFilesMethod processMethod = ProcessFilesMethod::Default;
@@ -487,7 +487,7 @@ void AssetImporterManager::ProcessCopyFiles()
 
 void AssetImporterManager::ProcessMoveFiles()
 {
-    int numberOfFiles = m_pathMap.size();
+    int numberOfFiles = static_cast<int>(m_pathMap.size());
     int numberOfProcessedFiles = 0;
 
     ProcessFilesMethod processMethod = ProcessFilesMethod::Default;
@@ -683,7 +683,7 @@ bool AssetImporterManager::Overwrite(QString relativePath, QString oldAbsolutePa
 
     while (!dataStream.atEnd())
     {
-        int bytesRead = dataStream.readRawData(buffer, bufferSize);
+        int bytesRead = static_cast<int>(dataStream.readRawData(buffer, bufferSize));
         out.writeRawData(buffer, bytesRead);
     }
 
@@ -740,7 +740,7 @@ bool AssetImporterManager::GetAndCheckAllFilesInFolder(QString path)
 
     // Get the index of the last sub folder name in the path
     QStringList directoryNameList = formattedPath.split('/');
-    int lastFolderIndex = directoryNameList.size() - 1;
+    int lastFolderIndex = static_cast<int>(directoryNameList.size()) - 1;
 
     QString pathToBeRelativeTo = directoryNameList.mid(0, lastFolderIndex).join('/');
