+++ Code/Tools/LuaIDE/Source/LUA/CodeCompletion/LUACompleter.cpp
@@ -33,6 +33,6 @@ namespace LUAEditor
     int Completer::GetCompletionPrefixTailLength()
     {
         auto result = completionPrefix().split(QRegularExpression(c_luaSplit));
-        return result.back().length();
+        return static_cast<int>(result.back().length());
     }
 }
