+++ Gems/AudioSystem/Code/Source/Editor/AudioControlsLoader.cpp
@@ -191,7 +191,7 @@ namespace AudioControls
     QStandardItem* CAudioControlsLoader::AddUniqueFolderPath(QStandardItem* parentItem, const QString& path)
     {
         QStringList folderNames = path.split(QRegularExpression("(\\\\|\\/)"), Qt::SkipEmptyParts);
-        const int size = folderNames.length();
+        const int size = static_cast<int>(folderNames.length());
         for (int i = 0; i < size; ++i)
         {
             if (!folderNames[i].isEmpty())
